Remove unnecessary items from the `ErrorCode` compat shim.
This type used to be the primary way to discover platform-specific error
numbers, but that purpose is now handled by the `freebsd-errno` and
`linux-errno` libraries. A shim `ErrorCode` struct was kept around to
reduce the size of the diff.

Some of the items in `ErrorCode` can be tidied up because they're no
longer referenced anywhere (`E2BIG`, `EINTR`, `ERANGE`). Others (the
`*_I32` items) are only referenced in `file_stream.rs` and would be
better placed in that file.

GitOrigin-RevId: 4ee85d8ec337688d1c9ea54b209d139c741c2417