Finish replacing `ErrorCode` with new `Error` struct.
This type is designed to simplify direct use of OS-specific error
numbers from the `linux-errno` or `freebsd-errno` libraries. Instead
of implementing `Into<u16>` for OS errors, which isn't possible for
FreeBSD, the `Error` type can implement `From` with a sentinal value
for error numbers that will be rejected by the client.

The `Error` struct also contains associated constants with OS-agnostic
names, which function similarly to the `std::io::ErrorKind` enum. This
simplifies server handlers that need to signal certain common error
conditions, or that need to return different errors according to the
target platform (such as `ENOATTR` vs `ENODATA` in xattr operations).

GitOrigin-RevId: de34d897e42669a73a3bce8cdd42ccc362312638