Add types for building the data parameter for `mount()`.
On Linux the `mount()` syscall traditionally receives FS-specific
options via the `void *data` parameter. For FUSE this parameter is
a C string containing comma-separated key-value pairs.

The new Linux-specific `MountOptions` struct represents options that
can be passed to the `mount()` call. It can be used to construct
a `MountData`, which itself can be used to obtain a `&[u8]` or `&CStr`
suitable for use as the `data` parameter.

Note that Linux's newer `fsmount` API uses a more structured encoding
of FS-specific options, and therefore doesn't need `MountData`.
FUSE server implementations based on `fsmount` would need to write
their own equivalent of `MountData`, but could still use `MountOptions`.

Note also that not all fields of `MountOptions` are passed via the
`data` parameter -- the `fs_type` and `source` options are passed as
separate `char*` parameters to `mount()`.

GitOrigin-RevId: ce55e636cfd993c5ad25a1487f839d45f919dafc