Move `UnknownRequest` to the `server` module.
It doesn't make sense for this type to be located with the known
operations, and `server` is the best place for it because a client
wouldn't have a need to parse requests with unknown opcodes.

The type's construction path is also changed to direct functions
`from_{cuse,fuse}_request`, taking a `{Cuse,Fuse}Request` respectively.

GitOrigin-RevId: ee5d53e3a245965bcf6088d1f57ca5a26747c326