Split module `server::basic` into `server::{cuse,fuse}_rpc`.
The previous arrangement was overly constraining on the design of
RPC-style CUSE/FUSE servers, because it positioned them as "basic"
compared to a low-level (opcode dispatch) or high-level (VFS) design.

The new modules have some duplicated API surface and will likely need
further cleanup to deal with the `SendReply` trait.

GitOrigin-RevId: ccbc8d12b2bd5486807996db97093caf279fd9a9