Merge remainder of `CuseConnection` into `cuse_rpc`.
The expanded `CuseServer` / `CuseServerBuilder` takes over the session
setup responsibility and socket ownership of `CuseConnection`.

The API of `cuse_rpc` can be simplified to remove non-CUSE symbols,
which dated from a time when the CUSE and FUSE server APIs were thought
to be potentially mergeable. While this might not be impossible, trying
to represent both CUSE and FUSE in the same module made the API
difficult to document. Separating them is an overall improvement.

Also includes some minor tweaks to the `CuseServer{,Builder}` types,
such as moving the optional `ServerHooks` into a `Box<>` to reduce
generic type parameters. The server hooks are primarily designed for
debugging, so there's little need to support them in a `no_std` build.

GitOrigin-RevId: c84bb320181c0e2edf541c18c6079a057e4b1fca