Move per-operation implementation modules to `operations`.
The `operations` module contains only code for the various FUSE
operations, and doesn't have any of the support code such as types or
helper macros. The more focused scope makes it easier to verify test
coverage, and to generate nice API docs.

To reduce the diff size the `protocol` module still exists, but is now
non-exported. In the future its content will be redistributed in the
`internal/` directory.

GitOrigin-RevId: 6cf9ba87a75e237385f05b3b30785e32c9b5ca1b