Build per-operation tests separately from the `fuse` library.
This lets Bazel reuse the same `fuse.rlib` for the tests, instead of
recompiling it with different `#[test]` sub-modules for each target.
It's also a good exercise of the public API, since the tests no longer
have access to `pub(crate)` symbols.

The `testutil` module is also extracted into a new `fuse_testutil`
library, along with a version of `fuse_kernel` that has all symbols set
to public visibility.

GitOrigin-RevId: ced35224ad237d78bc6dbd39b355d66ab1e027be