Simplify `readdir` responses to remove alignment handling.
Responses are now constructed using a `ReaddirEntriesWriter`, which
writes entries into a `&mut [u8]` provided by the user. There's no
alignment requirements for this buffer, and expectations about whether
it's allocated on the stack or heap.

Some of the low-level code for manipulating `fuse_dirent` is moved to
a new `internal::dirent` module.

GitOrigin-RevId: d470cb8d74652e549adcddcc67841712ae82ee1a