Make `decode` module public, and go back to decoder traits.
Having traits for decoding of FUSE and CUSE requests reduces the unique
API surface of the `*Request` structs, but the obvious trait names
of `{Cuse,Fuse}Request` conflict with the generic request containers of
the same name. Making the `decode` module public solves that problem,
because `server::FuseRequest` and `server::decode::FuseRequest` are
easily distinguishable.

This change also makes it practical to re-introduce `TryFrom`
implementations for the request structs without having to do too much
weird macro crafting.

GitOrigin-RevId: eb59016493ec6043ab2f57dc5817d80d875cc9e6