Add types for aligned buffers to `fuse::io`.
The `AlignedSlice` type is a marker/wrapper similar in purpose to the
old `fuse::io::Buffer` trait, but easier to work with. It provides a
safe `new(&[u8]) -> Option<Self>` constructor that verifies alignment
of a byte slice, and an unsafe `new_unchecked(&[u8]) -> Self` that
skips the verification.

The `ArrayBuffer` type is public again, but this time named
`MinReadBuffer` and with a tidier API.

GitOrigin-RevId: 8d644b282abf426237e2c6d1173e899fdd62c23d