Use pointer type-punning for `WriteRequest`.
This reduces the size of `WriteRequest` from 40 bytes to 16 (on x86_64),
allowing it to fit into registers instead of getting spilled to the
stack.

The `write_decode` benchmark shows the most improvement (as expected),
going from ~22.5ns to ~3.5ns (~85% reduction). The `write_dispatch` and
`write_end_to_end` benchmarks also see improvement (~30% and ~7.5%).

GitOrigin-RevId: f83335e650bf9d96ab54393cd3bfaa606e7ad509